avoid loading the same schema multiple times
This speeds up load_and_validate operations considerably for meta-schemas,
which contain a lot of references back to itself and we were going to disk
to reload the schema every time.

Also always register the schema at the location we loaded it from,
which may be different than the schema's $id (or it may be missing $id
entirely).  This is consistent with the spec at
https://tools.ietf.org/html/draft-handrews-json-schema-01#section-8.1.