fix resolution of $refs like "/foo/bar/baz"
These are perfectly valid uris on their own (as a filename or a path to be
passed to Mojo::UserAgent), or made absolute with a base like
"http://foo.com".

An important breaking side effect of this change is fragment-only $refs can no
longer be declared as "/foo/bar/baz" - this should have never been allowed
but some schemas in the wild may have done this because it worked until now.

Also fixed some inconsistent handling of base uris with respect to URL
escaping.